MODDCB-21-24-43: AWAITING_PICKUP-> ITEM_CHECKED_OUT + ITEM_CHECKED_OUT-> ITEM_CHECKED_IN + ITEM_CHECKED_IN->CLOSED: Implement DCB transaction status transition (Check-in at Pickup library, Pickup role)
