+++ src/main/java/org/folio/dcb/service/impl/BorrowingPickupLibraryServiceImpl.java
@@ -2,28 +2,17 @@ package org.folio.dcb.service.impl;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
-import org.folio.dcb.domain.dto.CirculationItemRequest;
 import org.folio.dcb.domain.dto.DcbTransaction;
 import org.folio.dcb.domain.dto.TransactionStatus;
 import org.folio.dcb.domain.dto.TransactionStatusResponse;
 import org.folio.dcb.domain.entity.TransactionEntity;
-import org.folio.dcb.repository.TransactionRepository;
 import org.folio.dcb.service.CirculationItemService;
-import org.folio.dcb.service.CirculationService;
 import org.folio.dcb.service.LibraryService;
 import org.folio.dcb.service.RequestService;
 import org.folio.dcb.service.UserService;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
 
-import java.util.UUID;
-
-import static org.folio.dcb.domain.dto.ItemStatus.NameEnum.AWAITING_PICKUP;
-import static org.folio.dcb.domain.dto.ItemStatus.NameEnum.CHECKED_OUT;
-import static org.folio.dcb.domain.dto.TransactionStatus.StatusEnum.CLOSED;
-import static org.folio.dcb.domain.dto.TransactionStatus.StatusEnum.CREATED;
-import static org.folio.dcb.domain.dto.TransactionStatus.StatusEnum.ITEM_CHECKED_IN;
-import static org.folio.dcb.domain.dto.TransactionStatus.StatusEnum.ITEM_CHECKED_OUT;
-import static org.folio.dcb.domain.dto.TransactionStatus.StatusEnum.OPEN;
 
 @Service("borrowingPickupLibraryService")
 @RequiredArgsConstructor
@@ -33,9 +22,8 @@ public class BorrowingPickupLibraryServiceImpl implements LibraryService {
   private final UserService userService;
   private final RequestService requestService;
   private final CirculationItemService circulationItemService;
-  private final TransactionRepository transactionRepository;
-  private final CirculationService circulationService;
-
+  @Qualifier("basicPickupLibraryService")
+  private final LibraryService basicPickupLibraryService;
 
   @Override
   public TransactionStatusResponse createCirculation(String dcbTransactionId, DcbTransaction dcbTransaction, String pickupServicePointId) {
@@ -55,45 +43,13 @@ public class BorrowingPickupLibraryServiceImpl implements LibraryService {
   }
 
   @Override
-  public void updateTransactionStatus(TransactionEntity dcbTransaction, TransactionStatus transactionStatus) {
-    log.debug("updateTransactionStatus:: Updating dcbTransaction {} to status {} ", dcbTransaction, transactionStatus);
-    var currentStatus = dcbTransaction.getStatus();
-    var requestedStatus = transactionStatus.getStatus();
-    if (CREATED == currentStatus && OPEN == requestedStatus) {
-      log.info("updateTransactionStatus:: Checking in item by barcode: {} ", dcbTransaction.getItemBarcode());
-      //Random UUID for servicePointId.
-      circulationService.checkInByBarcode(dcbTransaction, UUID.randomUUID().toString());
-      updateTransactionEntity(dcbTransaction, requestedStatus);
-    } else if (ITEM_CHECKED_IN == currentStatus && CLOSED == requestedStatus) {
-      updateTransactionEntity(dcbTransaction, requestedStatus);
-    } else {
-      String errorMessage = String.format("updateTransactionStatus:: status update from %s to %s is not implemented", currentStatus, requestedStatus);
-      log.warn(errorMessage);
-      throw new IllegalArgumentException(errorMessage);
-    }
+  public void updateStatusByTransactionEntity(TransactionEntity transactionEntity) {
+    basicPickupLibraryService.updateStatusByTransactionEntity(transactionEntity);
   }
 
   @Override
-  public void updateStatusByTransactionEntity(TransactionEntity transactionEntity) {
-    log.debug("updateTransactionStatus:: Received checkIn event for itemId: {}", transactionEntity.getItemId());
-    CirculationItemRequest circulationItemRequest = circulationItemService.fetchItemById(transactionEntity.getItemId());
-    if (OPEN == transactionEntity.getStatus() && AWAITING_PICKUP == circulationItemRequest.getStatus().getName()) {
-      updateTransactionEntity(transactionEntity, TransactionStatus.StatusEnum.AWAITING_PICKUP);
-    } else if (TransactionStatus.StatusEnum.AWAITING_PICKUP == transactionEntity.getStatus() && CHECKED_OUT == circulationItemRequest.getStatus().getName()) {
-      updateTransactionEntity(transactionEntity, ITEM_CHECKED_OUT);
-    } else if(ITEM_CHECKED_OUT == transactionEntity.getStatus()){
-      log.info("updateStatusByTransactionEntity:: Updated item status from {} to {}",
-        transactionEntity.getStatus().getValue(), TransactionStatus.StatusEnum.ITEM_CHECKED_IN.getValue());
-      updateTransactionEntity(transactionEntity, TransactionStatus.StatusEnum.ITEM_CHECKED_IN);
-    } else {
-      log.info("updateStatusByTransactionEntity:: Item status is {}. So status of transaction is not updated",
-        circulationItemRequest.getStatus().getName());
-    }
+  public void updateTransactionStatus(TransactionEntity dcbTransaction, TransactionStatus transactionStatus) {
+    basicPickupLibraryService.updateTransactionStatus(dcbTransaction, transactionStatus);
   }
 
-  private void updateTransactionEntity(TransactionEntity transactionEntity, TransactionStatus.StatusEnum transactionStatusEnum) {
-    log.info("updateTransactionEntity:: updating transaction entity from {} to {}", transactionEntity.getStatus(), transactionStatusEnum);
-    transactionEntity.setStatus(transactionStatusEnum);
-    transactionRepository.save(transactionEntity);
-  }
 }
