--- src/test/java/org/folio/dcb/service/BorrowingLibraryServiceTest.java
@@ -1,78 +1,0 @@
-package org.folio.dcb.service;
-
-import org.folio.dcb.domain.dto.TransactionStatus;
-import org.folio.dcb.repository.TransactionRepository;
-import org.folio.dcb.service.impl.BorrowingPickupLibraryServiceImpl;
-import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.InjectMocks;
-import org.mockito.Mock;
-import org.folio.dcb.domain.dto.ItemStatus;
-import org.folio.dcb.domain.dto.CirculationItemRequest;
-
-import static org.folio.dcb.domain.dto.DcbTransaction.RoleEnum.BORROWING_PICKUP;
-import static org.folio.dcb.domain.dto.TransactionStatus.StatusEnum.CLOSED;
-import static org.junit.jupiter.api.Assertions.assertThrows;
-
-import org.mockito.Mockito;
-import org.mockito.junit.jupiter.MockitoExtension;
-
-import static org.folio.dcb.utils.EntityUtils.createTransactionEntity;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.Mockito.*;
-
-@ExtendWith(MockitoExtension.class)
-class BorrowingLibraryServiceTest {
-  @InjectMocks
-  private BorrowingPickupLibraryServiceImpl borrowingPickupLibraryService;
-  @Mock
-  private TransactionRepository transactionRepository;
-  @Mock
-  private CirculationService circulationService;
-  @Mock
-  private CirculationItemService circulationItemService;
-
-  @Test
-  void updateTransactionTestFromCreatedToOpen() {
-    var transactionEntity = createTransactionEntity();
-    transactionEntity.setStatus(TransactionStatus.StatusEnum.CREATED);
-    transactionEntity.setRole(BORROWING_PICKUP);
-    doNothing().when(circulationService).checkInByBarcode(any(), any());
-
-    borrowingPickupLibraryService.updateTransactionStatus(transactionEntity, TransactionStatus.builder().status(TransactionStatus.StatusEnum.OPEN).build());
-    verify(transactionRepository, times(1)).save(transactionEntity);
-    verify(circulationService, timeout(1)).checkInByBarcode(any(), any());
-  }
-
-  @Test
-  void updateTransactionErrorTest() {
-    var transactionEntity = createTransactionEntity();
-    transactionEntity.setStatus(TransactionStatus.StatusEnum.CREATED);
-    transactionEntity.setRole(BORROWING_PICKUP);
-    TransactionStatus transactionStatus = TransactionStatus.builder().status(TransactionStatus.StatusEnum.AWAITING_PICKUP).build();
-    assertThrows(IllegalArgumentException.class, () -> borrowingPickupLibraryService.updateTransactionStatus(transactionEntity, transactionStatus));
-  }
-
-  @Test
-  void updateTransactionTestFromCheckedOutToCheckedIns() {
-    var transactionEntity = createTransactionEntity();
-    transactionEntity.setStatus(TransactionStatus.StatusEnum.ITEM_CHECKED_OUT);
-    transactionEntity.setRole(BORROWING_PICKUP);
-    when(circulationItemService.fetchItemById(any())).thenReturn(CirculationItemRequest.builder().status(
-      ItemStatus.builder().name(ItemStatus.NameEnum.AVAILABLE).build()).build());
-    borrowingPickupLibraryService.updateStatusByTransactionEntity(transactionEntity);
-    Mockito.verify(transactionRepository, times(1)).save(transactionEntity);
-  }
-
-  @Test
-  void updateTransactionStatusFromItemCheckedInToClosedTest() {
-    var transactionEntity = createTransactionEntity();
-    transactionEntity.setStatus(TransactionStatus.StatusEnum.ITEM_CHECKED_IN);
-    transactionEntity.setRole(BORROWING_PICKUP);
-
-    borrowingPickupLibraryService.updateTransactionStatus(transactionEntity, TransactionStatus.builder().status(CLOSED).build());
-
-    Assertions.assertEquals(CLOSED, transactionEntity.getStatus());
-  }
-}
